+++ src/api/robot_environments/interfaces/basic/index.php
@@ -61,6 +61,20 @@ class basic
 <html>
 <head>
 <?php $re->create_head() // grab the header information ?>
+<?php if (isset($_SERVER['HTTPS'])) {?>
+<script type="text/javascript"
+    src="https://robotsfor.me/cdn/EventEmitter2/0.4.11/eventemitter2.js">
+</script>
+<script type="text/javascript"
+    src="https://robotsfor.me/cdn/roslibjs/r5/roslib.min.js"></script>
+<script type="text/javascript"
+    src="https://robotsfor.me/cdn/mjpegcanvasjs/r1/mjpegcanvas.min.js">
+</script>
+<script type="text/javascript"
+  src="https://robotsfor.me/cdn/keyboardteleopjs/r1/keyboardteleop.min.js">
+</script>
+<?php 
+} else {?>
 <script type="text/javascript"
     src="http://cdn.robotwebtools.org/EventEmitter2/0.4.11/eventemitter2.js">
 </script>
@@ -72,6 +86,8 @@ class basic
 <script type="text/javascript"
   src="http://cdn.robotwebtools.org/keyboardteleopjs/r1/keyboardteleop.min.js">
 </script>
+<?php 
+} ?>
 <title>Basic Teleop Interface</title>
 <script type="text/javascript">
   //connect to ROS